refactor(approval-gate): extract delivery.rs from lib.rs
Sixth step of breaking lib.rs into focused modules. This commit
lifts the delivery-tracking RPCs into their own module.

Moved (lib.rs -> delivery.rs):
- handle_list_pending()
- handle_list_undelivered()
- handle_ack_delivered()
- handle_consume_undelivered()
- handle_flush_delivered()
- handle_sweep_session()
- LIST_UNDELIVERED_DEFAULT_LIMIT

All public items stay re-exported from the crate root.
Pure move, no behavior change. 141 approval-gate tests pass.
